requirements.txt
python-telegram-bot==20.7
sentence-transformers
faiss-cpu
ollama
pypdf
numpy
